+++ apps/hibiki/lib/hibiki/command/tl.ex
@@ -14,8 +14,14 @@ defmodule Hibiki.Command.Tl do
   def handle(%{"query" => query}, _ctx) do
     url = Hibiki.Config.deepl_proxy() <> "/translate/tl"
 
+    http_config = [
+      hackney: [pool: :tl],
+      recv_timeout: 30_000,
+      timeout: 30_000
+    ]
+
     with {:ok, body} <- Jason.encode(%{"src" => query}),
-         {:ok, %HTTPoison.Response{body: body}} <- HTTPoison.post(url, body),
+         {:ok, %HTTPoison.Response{body: body}} <- HTTPoison.post(url, body, [], http_config),
          {:ok, %{"result" => result, "src_lang" => src_lang}} <- Jason.decode(body) do
       {:reply,
        %LineSdk.Model.TextMessage{