+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 ### Changed
 - Changed #tag# to ignore if the last char inside is space (like #tag #tag)
 - Removed unnecessary logging in tenshi provider
+- Add httpoison options to post call in !tl